Add files via upload
We modeled an SVM radial classification machine learning algorithm to determine the ruptured and unruptured risk of saccular cerebral aneurysms using 60 samples with 6 predictors as the gender, the age, the Womersley number, the Time-Averaged Wall Shear Stress (TAWSS), the Aspect Ratio (AR) and the bottleneck of the aneurysms, considering real cases of patients. We reconstructed computationally each geometry from an angiography image to realize a CFD simulations, where the TAWSS was computed by CFD analysis. A cross validation method was used in the training sample to validate the classification model, getting an accuracy of 92.86% in the test sample. This result may be used to help in medical decisions to avoid a complicated operation when the probability of rupture is low.