Add files via upload
Fluid-mechanical and morphological parameters are recognized as major factors in the rupture risk of human aneurysms. On the other hand, it is well known that a lot of machine learning tools are available to study a variety of problems in many fields. In this work, fluid–structure interaction (FSI) simulations were carried out to examine a database of 60 real saccular cerebral aneurysms (30 ruptured and 30 unruptured) using reconstructions by angiography images. With the results of the simulations and geometric analyses, we studied the analysis of variance (ANOVA) statistic test in many variables and we obtained that aspect ratio (AR), bottleneck factor (BNF), maximum height of the aneurysms (MH), relative residence time (RRT), Womersley number (WN) and Von-Mises strain (VMS) are statically significant and good predictors for the models. In consequence, these ones were used in five machine learning algorithms to determine the rupture risk predictions of the aneurysms, where the adaptative boosting (AdaBoost) was calculated with the highest area under the curve (AUC) in the receiver operating characteristic (ROC) curve (AUC 0.944).